# backend/docs/openapi-0.4.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ThermoTrace API",
    "description": "API do ThermoTrace — monitoramento térmico de remessas por etiqueta NFC.\n\nVersão 0.4: acesso, perfis, cargas e recepção idempotente de evidências. Declarações de início/checkpoint/final preservadas com cadeia de integridade. Receber uma evidência não confirma o hardware nem libera uma carga. Fila Android e decodificação no servidor estão em implementação.",
    "version": "0.4.0"
  },
  "paths": {
...
    "/api/v1/volumes/{volume_id}/sessoes": {
      "get": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Sessoes Do Volume",
        "operationId": "sessoes_do_volume_api_v1_volumes__volume_id__sessoes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Volume Id"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Iniciar",
        "operationId": "iniciar_api_v1_volumes__volume_id__sessoes_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "volume_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Volume Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IniciarSessao"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReciboLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/dispositivos": {
      "post": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Dispositivo",
        "operationId": "dispositivo_api_v1_dispositivos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrarDispositivo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/etiquetas/localizar": {
      "get": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Etiqueta",
        "operationId": "etiqueta_api_v1_etiquetas_localizar_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(?:[0-9A-F]{2}){4,16}$",
              "title": "Uid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessoes/{sessao_id}/leituras": {
      "post": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Leitura",
        "operationId": "leitura_api_v1_sessoes__sessao_id__leituras_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sessao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sessao Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 16,
              "maxLength": 160,
              "pattern": "^[A-Za-z0-9._:-]+$",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceberLeitura"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReciboLeitura"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Listar",
        "operationId": "listar_api_v1_sessoes__sessao_id__leituras_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sessao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sessao Id"
            }
          },
          {
            "name": "apos_ordem",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Apos Ordem"
            }
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReciboLeitura"
                  },
                  "title": "Response Listar Api V1 Sessoes  Sessao Id  Leituras Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessoes/{sessao_id}/integridade": {
      "get": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Integridade",
        "operationId": "integridade_api_v1_sessoes__sessao_id__integridade_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sessao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sessao Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sessoes/{sessao_id}/leituras/{leitura_id}": {
      "get": {
        "tags": [
          "Recepção de evidências"
        ],
        "summary": "Evidencia",
        "operationId": "evidencia_api_v1_sessoes__sessao_id__leituras__leitura_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sessao_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Sessao Id"
            }
          },
          {
            "name": "leitura_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Leitura Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Evidencia": {
        "properties": {
          "evento_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evento Id"
          },
          "dispositivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispositivo Id"
          },
          "lida_em": {
            "type": "string",
            "format": "date-time",
            "title": "Lida Em"
          },
          "uid_canonico": {
            "type": "string",
            "pattern": "^(?:[0-9A-F]{2}){4,16}$",
            "title": "Uid Canonico"
          },
          "versao_sdk": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao Sdk"
          },
          "versao_decodificador": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao Decodificador"
          },
          "resposta_bruta": {
            "items": {
              "type": "string",
              "maxLength": 131072
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Resposta Bruta"
          },
          "origem": {
            "type": "string",
            "enum": [
              "declaracao_android",
              "simulacao"
            ],
            "title": "Origem"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evento_id",
          "dispositivo_id",
          "lida_em",
          "uid_canonico",
          "versao_sdk",
          "versao_decodificador",
          "resposta_bruta",
          "origem"
        ],
        "title": "Evidencia"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IniciarSessao": {
        "properties": {
          "sessao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sessao Id"
          },
          "etiqueta_id": {
            "type": "string",
            "format": "uuid",
            "title": "Etiqueta Id"
          },
          "evidencia": {
            "$ref": "#/components/schemas/Evidencia"
          },
          "epoch_inicio_etiqueta": {
            "type": "integer",
            "maximum": 4133980800.0,
            "minimum": 946684800.0,
            "title": "Epoch Inicio Etiqueta"
          },
          "delay_minutos": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 0.0,
            "title": "Delay Minutos",
            "default": 0
          },
          "intervalo_segundos": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Intervalo Segundos"
          },
          "quantidade_planejada": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Quantidade Planejada"
          },
          "min_configurado_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,7}0*$)\\d{0,4}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Min Configurado C"
          },
          "max_configurado_c": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,4}|(?=[\\d.]{1,7}0*$)\\d{0,4}\\.\\d{0,2}0*$)"
              }
            ],
            "title": "Max Configurado C"
          },
          "ativacao_confirmada": {
            "type": "boolean",
            "const": true,
            "title": "Ativacao Confirmada"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "sessao_id",
          "etiqueta_id",
          "evidencia",
          "epoch_inicio_etiqueta",
          "intervalo_segundos",
          "quantidade_planejada",
          "min_configurado_c",
          "max_configurado_c",
          "ativacao_confirmada"
        ],
        "title": "IniciarSessao"
      },
...
      "ReceberLeitura": {
        "properties": {
          "evento_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evento Id"
          },
          "dispositivo_id": {
            "type": "string",
            "format": "uuid",
            "title": "Dispositivo Id"
          },
          "lida_em": {
            "type": "string",
            "format": "date-time",
            "title": "Lida Em"
          },
          "uid_canonico": {
            "type": "string",
            "pattern": "^(?:[0-9A-F]{2}){4,16}$",
            "title": "Uid Canonico"
          },
          "versao_sdk": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao Sdk"
          },
          "versao_decodificador": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao Decodificador"
          },
          "resposta_bruta": {
            "items": {
              "type": "string",
              "maxLength": 131072
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 1,
            "title": "Resposta Bruta"
          },
          "origem": {
            "type": "string",
            "enum": [
              "declaracao_android",
              "simulacao"
            ],
            "title": "Origem"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "checkpoint",
              "final"
            ],
            "title": "Tipo"
          },
          "epoch_inicio_etiqueta": {
            "type": "integer",
            "maximum": 4133980800.0,
            "minimum": 946684800.0,
            "title": "Epoch Inicio Etiqueta"
          },
          "intervalo_relatado_s": {
            "type": "integer",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Intervalo Relatado S"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "evento_id",
          "dispositivo_id",
          "lida_em",
          "uid_canonico",
          "versao_sdk",
          "versao_decodificador",
          "resposta_bruta",
          "origem",
          "tipo",
          "epoch_inicio_etiqueta",
          "intervalo_relatado_s"
        ],
        "title": "ReceberLeitura"
      },
      "ReciboLeitura": {
        "properties": {
          "leitura_id": {
            "type": "string",
            "format": "uuid",
            "title": "Leitura Id"
          },
          "sessao_id": {
            "type": "string",
            "format": "uuid",
            "title": "Sessao Id"
          },
          "evento_id": {
            "type": "string",
            "format": "uuid",
            "title": "Evento Id"
          },
          "empresa_id": {
            "type": "string",
            "format": "uuid",
            "title": "Empresa Id"
          },
          "volume_id": {
            "type": "string",
            "format": "uuid",
            "title": "Volume Id"
          },
          "remessa_id": {
            "type": "string",
            "format": "uuid",
            "title": "Remessa Id"
          },
          "tipo": {
            "type": "string",
            "title": "Tipo"
          },
          "origem": {
            "type": "string",
            "enum": [
              "declaracao_android",
              "simulacao"
            ],
            "title": "Origem"
          },
          "ordem_recebimento": {
            "type": "integer",
            "title": "Ordem Recebimento"
          },
          "recebida_em_servidor": {
            "type": "string",
            "format": "date-time",
            "title": "Recebida Em Servidor"
          },
          "hash_payload": {
            "type": "string",
            "title": "Hash Payload"
          },
          "hash_anterior": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hash Anterior"
          },
          "hash_encadeado": {
            "type": "string",
            "title": "Hash Encadeado"
          },
          "algoritmo_integridade": {
            "type": "string",
            "const": "tt-evidencia-1",
            "title": "Algoritmo Integridade",
            "default": "tt-evidencia-1"
          },
          "situacao": {
            "type": "string",
            "const": "evidencia_recebida_sem_validacao_fisica",
            "title": "Situacao",
            "default": "evidencia_recebida_sem_validacao_fisica"
          },
          "stop_fisico": {
            "type": "string",
            "const": "nao_confirmado_pelo_servidor",
            "title": "Stop Fisico",
            "default": "nao_confirmado_pelo_servidor"
          }
        },
        "type": "object",
        "required": [
          "leitura_id",
          "sessao_id",
          "evento_id",
          "empresa_id",
          "volume_id",
          "remessa_id",
          "tipo",
          "origem",
          "ordem_recebimento",
          "recebida_em_servidor",
          "hash_payload",
          "hash_anterior",
          "hash_encadeado"
        ],
        "title": "ReciboLeitura"
      },
      "RegistrarDispositivo": {
        "properties": {
          "install_id": {
            "type": "string",
            "format": "uuid",
            "title": "Install Id"
          },
          "modelo": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Modelo"
          },
          "versao_so": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao So"
          },
          "versao_app": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Versao App"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "install_id",
          "modelo",
          "versao_so",
          "versao_app"
        ],
        "title": "RegistrarDispositivo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}